use-social-auth/django-social/social_auth_app/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-allauth==0.54.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
oauthlib==3.2.2
psutil==5.9.5
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.0
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.4.0
